% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_files.R
\name{add_local_file}
\alias{add_local_file}
\title{Add files to local SeqPlots installation}
\usage{
add_local_file(x, name = deparse(substitute(x)), file_genome = "ce11",
  file_user = "user", file_comment = "",
  root = file.path(path.expand("~"), "SeqPlots_data"))
}
\arguments{
\item{x}{file path, GRanges, SimpleRleList or BigWigFile object}

\item{name}{neme of the file}

\item{file_genome}{genome version}

\item{file_user}{user name}

\item{file_comment}{comment}

\item{root}{local installation directory}
}
\description{
Add files to local SeqPlots installation
}
