% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployServerInstance.R
\name{deployServerInstance}
\alias{deployServerInstance}
\title{Prepares the instance of SeqPlots, that can be deployed to shinyapps.io}
\usage{
deployServerInstance(server = getwd(), data = file.path(server, "DATA"))
}
\arguments{
\item{server}{The directory where server instance should be created.}

\item{data}{the data storige directory, should be a sub-dir of "server" for 
shinyapps}
}
\value{
\code{NULL}
}
\description{
See http://shiny.rstudio.com/articles/shinyapps.html for further info.
}
\keyword{internal}

