% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPlotSetArray}
\alias{getPlotSetArray}
\title{Process genomic signal}
\usage{
getPlotSetArray(tracks, features, refgenome, bin = 10L, rm0 = FALSE,
  ignore_strand = FALSE, x1 = 2000L, x2 = 2000L, xm = 1000L,
  type = "pf", add_heatmap = TRUE, cat3 = message, cat4 = message,
  con = NULL)
}
\arguments{
\item{tracks}{list or vector of track paths (BigWig) and/or motif
setup structers (as \code{\link{list}})}

\item{features}{vector of feature file paths (BED or GFF)}

\item{refgenome}{the UCSC cole of refference genome, e.g. 'hg19' hor
Homo sapiens}

\item{bin}{binning window size in base pairs, default 1L}

\item{rm0}{remove 0 from mean/median calculation, default FALSE}

\item{x1}{upsterem distance in base pairs, default 500L}

\item{x2}{downsteram distance in base pairs, default 2000L}

\item{xm}{anchored distance in base pairs, default 1000L}

\item{type}{the type of the calculation, default, 'Point Features'}

\item{add_heatmap}{return the heatmap data,  default FALSE}

\item{cat3}{function, that sends 1st level info to web interface, handeld by
"message" in console mode}

\item{cat4}{function, that sends 2nd level info to web interface, handeld by
"message" in console mode}

\item{con}{connection to SQlite database storing genome information for files}
}
\value{
nested list: OUT[[FEATURE]][[TRACK/MOTIF]][[VALUE]]
}
\description{
This function extracts and processes genomic signal from tracks and/or
motif data, calculates statistics and presents the data in nested list format.
\code{\link{transform}} but for reordering a data frame by its columns.
This saves a lot of typing!
}
\examples{
\dontrun{
getPlotSetArray(c('track1.bw', 'track2.bw'), c('peaks.bed', 'TSS.gff'))
}
}

