% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getREF}
\alias{getREF}
\title{Get reference genome}
\usage{
getREF(genome)
}
\arguments{
\item{genome}{The filename of FASTA file or genome code for BSgenome}
}
\value{
\code{DNAStringSet}
}
\description{
Get reference genome
}
\keyword{internal}
