% Generated by roxygen2 (4.0.1): do not edit by hand
\name{heatmapPlotWrapper}
\alias{heatmapPlotWrapper}
\title{Wrapper function, plotting the heatmap}
\usage{
heatmapPlotWrapper(MAT, axhline = NULL, titles = rep("", length(MAT)),
  bins = 1:(ncol(MAT[[1]])/length(MAT)), lfs = 12, afs = 12, lgfs = 12,
  xlabel = "xlab", Leg = TRUE, autoscale = TRUE, zmin = 0, zmax = 10,
  ln.v = TRUE, e = NULL, xlim = NULL, ylabel = "", s = 0.01,
  indi = TRUE, o_min = NA, o_max = NA, colvec = NULL,
  colorspace = NULL, poinsize = 12)
}
\description{
This function is package internal and should not be executed directly
by users.
}
\keyword{internal}

