% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{import.narrowPeak}
\alias{import.narrowPeak}
\title{Import function for narrowPeak format}
\usage{
import.narrowPeak(x)
}
\arguments{
\item{x}{}
}
\value{
GenomicRanges
}
\description{
Import function for narrowPeak format
}
