% Generated by roxygen2 (4.0.2): do not edit by hand
\name{num2bp}
\alias{num2bp}
\title{Converts numeric values to labels with base pairs units,
i.e bp, kb, Mb, Gb, etc.}
\usage{
num2bp(n)
}
\arguments{
\item{n}{The numeric value, that will be converted to label.}
}
\value{
\code{character}
}
\description{
Converts numeric values to labels with base pairs units,
i.e bp, kb, Mb, Gb, etc.
}
\keyword{internal}

