% Generated by roxygen2 (4.0.2): do not edit by hand
\name{num2sci}
\alias{num2sci}
\title{Converts numeric values labels with metric system suffix, i.e k, M, G, etc.}
\usage{
num2sci(n)
}
\arguments{
\item{n}{The numeric value, that will be converted to label.}
}
\value{
\code{character}
}
\description{
Converts numeric values labels with metric system suffix, i.e k, M, G, etc.
}
\keyword{internal}

