% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotAverage}
\alias{plotAverage}
\alias{plotAverage,PlotSetArray-method}
\alias{plotAverage,PlotSetList-method}
\alias{plotAverage,PlotSetPair-method}
\alias{plotAverage,list-method}
\title{Create the average plot}
\usage{
plotAverage(plotset, ...)

\S4method{plotAverage}{list}(plotset, keepratio = FALSE, ord = NULL,
  labels = NULL, ...)

\S4method{plotAverage}{PlotSetPair}(plotset, ...)

\S4method{plotAverage}{PlotSetList}(plotset, ...)

\S4method{plotAverage}{PlotSetArray}(plotset, ...)
}
\arguments{
\item{plotset}{The dataset to plot - can be \code{\link{PlotSetArray}},
\code{\link{PlotSetLits}}, \code{\link{PlotSetPair}} or
properly formated \code{\link[base]{list}}}

\item{main}{an overall title for the plot: see title.}

\item{xlab}{a title for the x axis: see title.}

\item{ylab}{a title for the y axis: see title.}

\item{plotScale}{scale the avilable data before ploting, can be "linear" (do not scale), "log2" or "zscore"}

\item{xlim}{the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed and leads to a ‘reversed axis’.
The default value, NULL, indicates that the whole range present in \code{plotset} will be plotted.}

\item{ylim}{the y limits (y1, y2) of the plot. Note that x1 > x2 is allowed and leads to a ‘reversed axis’.
The default value, NULL, indicates that the range will be auto calculated including space for error estimates.}

\item{type}{If set to "legend" only the legend/key will be plotted.}

\item{error.estimates}{Indicates if error estimates are plotted, default: TRUE}

\item{legend}{Indicates if key descibing the PlotSetPairs is shown, default: TRUE}

\item{legend_ex}{Indicates if key descibing error estimates is shown, default: FALSE}

\item{legend_pos}{The position of main key, default: 'topright'}

\item{legend_ext_pos}{The position of error estimates key,  default: 'topleft'}

\item{cex.axis}{Axis numbers font size in points, default: 14}

\item{cex.lab}{Axis labels labes font size in points, default: 16}

\item{cex.main}{Main title font size in points, default: 20}

\item{cex.legend}{Keys labels font size in points, default: 10}

\item{ln.v}{=TRUE,}

\item{ln.h}{=NULL,}

\item{colvec}{=NULL,}

\item{poinsize}{=12,}

\item{postproc}{=function(){}}

\item{keepratio}{}

\item{ord}{}

\item{labels}{}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\description{
Create the average plot
}
\details{
Soem details.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Method for signature plotset='list'

\item \code{PlotSetPair}: Method for signature plotset='PlotSetPair'

\item \code{PlotSetList}: Method for signature plotset='PlotSetList'

\item \code{PlotSetArray}: Method for signature plotset='PlotSetArray'
}}
\examples{
\dontrun{
## Run BEADS for BSgenome package, the reference genome package have to be installed prior to running this example
# source("http://bioconductor.org/biocLite.R")
# biocLite("BSgenome.Celegans.UCSC.ce10")
# library(BSgenome.Celegans.UCSC.ce10)
beads(sample_bam, SummedInput_bw, map_bw, genome='ce10')

## Run BEADS for all BAM files in the directory
#lapply(dir(pattern='bam$'), beads, control=input, mappability=map_bw, genome=ref_fa)
}
}
\author{
Przemyslaw Stempor
}
\keyword{internal}

