% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap with optional clustering}
\usage{
plotHeatmap(pl, title = "", labels = NA, legend = TRUE,
  keepratio = FALSE, ord = 1:length(pl), scale_signal = "no",
  sortrows = FALSE, clusters = 5L, clstmethod = "kmeans",
  include = rep(TRUE, length(pl)), ssomt1 = 2L, ssomt2 = 2L,
  title_font_size = 16, ...)
}
\arguments{
\item{tracks}{list or vector of track paths (BigWig) and/or motif
setup structers (as \code{\link{list}})}

\item{features}{vector of feature file paths (BED or GFF)}

\item{bin}{binning window size in base pairs, default 1L}

\item{rm0}{remove 0 from mean/median calculation, default FALSE}

\item{x1}{upsterem distance in base pairs, default 500L}

\item{x2}{downsteram distance in base pairs, default 2000L}

\item{xm}{anchored distance in base pairs, default 1000L}

\item{type}{the type of the calculation, default, 'Point Features'}

\item{add_heatmap}{return the heatmap data,  default FALSE}

\item{cat3}{function, that sends 1st level info to web interface}

\item{cat4}{function, that sends 2nd level info to web interface}

\item{con}{connection to SQlite database storing genome information for files}
}
\value{
nested list: OUT[[FEATURE]][[TRACK/MOTIF]][[VALUE]]
}
\description{
This function extracts and processes genomic signal from tracks and/or
motif data, calculates statistics and presents the data in nested list format.
\code{\link{transform}} but for reordering a data frame by its columns.
This saves a lot of typing!
}

