% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMext}
\alias{plotMext}
\title{Wrapper function, plotting the average plot}
\usage{
plotMext(INPUTS, xlim = NULL, ylim = NULL, main = NULL, xlab = "",
  ylab = "", plotScale = "linear", type = "full",
  error.estimates = TRUE, legend = TRUE, legend_ext = FALSE,
  legend_pos = "topright", legend_ext_pos = "topleft", cex.axis = 14,
  cex.lab = 16, cex.main = 20, cex.legend = 10, ln.v = TRUE,
  ln.h = NULL, colvec = NULL, poinsize = 12, postproc = function() { },
  ...)
}
\description{
This function is package internal and should not be executed directly
by users.
}
\keyword{internal}

