% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{PlotSetArray-class}
\alias{PlotSetArray}
\alias{PlotSetArray-class}
\title{PlotSetArray Reference Class}
\description{
PlotSetArray Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}

