% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{PlotSetList-class}
\alias{PlotSetList}
\alias{PlotSetList-class}
\title{PlotSetList Reference Class}
\description{
PlotSetList Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}

