% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{PlotSetPair-class}
\alias{PlotSetPair}
\alias{PlotSetPair-class}
\title{PlotSetPair Reference Class}
\description{
PlotSetPair Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a list holding the plot data}

\item{\code{annotations}}{list of annotations}
}}

