% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run}
\alias{run}
\title{SeqPlots innitiation}
\usage{
run(root = file.path(path.expand("~"), "SeqPlots_data"), debug = FALSE, ...)
}
\arguments{
\item{root}{the path to data directory, it will be created if not existing}

\item{debug}{run the SeqPlots in debyug mode, i.e. with web interface R console}

\item{...}{arguments sent to \code{\link[shiny]{runApp}} function}
}
\value{
Normally returns nothing (NULL), returns an error if one accrued.
Usage messages are shown in R console.
}
\description{
This function is called to initaite SeqPlots and call web browser graphical user interface.
}
\details{
The default home dir is ~/SeqPlots data
}
\examples{
\dontrun{
run()
}
}
\author{
Przemyslaw Stempor
}
\references{
https://bitbucket.org/przemol/seqplots/wiki/Home
}
\keyword{seqplots}

