% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{seqplots-generic}
\alias{[,PlotSetArray,ANY,missing,ANY-method}
\alias{[,PlotSetArray,ANY,vector,ANY-method}
\alias{[,PlotSetList,ANY,ANY,ANY-method}
\alias{[[,PlotSetArray,ANY,ANY-method}
\alias{[[,PlotSetList,ANY,ANY-method}
\alias{seqplots-generic}
\alias{unlist,PlotSetArray-method}
\title{Generic method for SeqPlots package}
\usage{
\S4method{[}{PlotSetList,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{PlotSetList,ANY,ANY}(x, i, j, ...)

\S4method{[}{PlotSetArray,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{PlotSetArray,ANY,vector,ANY}(x, i, j)

\S4method{[[}{PlotSetArray,ANY,ANY}(x, i, j, ...)

\S4method{unlist}{PlotSetArray}(x)
}
\description{
Generic method for SeqPlots package
}
\seealso{
Extract
}

