% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{seqplots}
\alias{seqplots}
\alias{seqplots-package}
\title{SeqPlots - An interactive tool for visualizing NGS signals and sequence
motif densities along genomic features using average plots and heatmaps.}
\description{
SeqPlots is a tool for plotting next generation sequencing (NGS) based
experiments' signal tracks, e.g. reads coverage from ChIP-seq, RNA-seq
and DNA accessibility assays like DNase-seq and MNase-seq, over user
specified genomic features, e.g. promoters, gene bodies, etc. It can
also calculate sequence motif density profiles from reference genome.
The data are visualized as average signal profile plot, with error estimates
(standard error and 95% confidence interval) shown as fields, or as series
of heatmaps that can be sorted and clustered using hierarchical clustering,
k-means algorithm and self organising maps. Plots can be prepared using R
programming language or web browser based graphical user interface (GUI)
implemented using Shiny framework. The dual-purpose implementation allows
running the software locally on desktop or deploying it on server.
SeqPlots is useful for both for exploratory data analyses and preparing
replicable, publication quality plots. Other features of the software
include collaboration and data sharing capabilities, as well as ability
to store pre-calculated result matrixes, that combine many sequencing
experiments and in-silico generated tracks with multiple different features.
These binaries can be further used to generate new combination plots on fly,
run automated batch operations or share with colleagues, who can adjust
their plotting parameters without loading actual tracks and recalculating
numeric values. SeqPlots relays on Bioconductor packages, mainly on
rtracklayer for data input and BSgenome packages for reference genome
sequence and annotations.
}
\author{
Przemyslaw Stempor
}
\references{
https://github.com/Przemol/seqplots
}
\keyword{CHiP-seq}
\keyword{genomics}
\keyword{plotting}
\keyword{sequencing}

