% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPlotSetArray}
\alias{getPlotSetArray}
\title{Process genomic signal}
\usage{
getPlotSetArray(tracks, features, refgenome, bin = 10L, rm0 = FALSE,
  ignore_strand = FALSE, xmin = 2000L, xmax = 2000L, xanchored = 1000L,
  type = "pf", add_heatmap = TRUE)
}
\arguments{
\item{tracks}{list or vector of track paths (BigWig) and/or motif
setup structers (as \code{\link{list}})}

\item{features}{vector of feature file paths (BED or GFF)}

\item{refgenome}{the UCSC cole of refference genome, e.g. 'hg19' hor
Homo sapiens}

\item{bin}{binning window size in base pairs, defaults t 1L}

\item{rm0}{remove 0 from mean/median calculation, defaults t FALSE}

\item{ignore_strand}{If TRUE the strand is ignored, that is
 all regardless og annotatin in GFF/BED file all strands are treated
 as undetrmined ("*"), defauls to FALSE}

\item{xmin}{upsterem distance in base pairs, defaults to 200L}

\item{xmax}{downsteram distance in base pairs, defaults to 2000L}

\item{xanchored}{distance in base pairs, defaults to 1000L}

\item{type}{the type of the calculation, "pf" for point features (default),
"mf" for midpoint features and 'af' for anchored features}

\item{add_heatmap}{return the heatmap data, defauls to TRUE}
}
\value{
\code{\link{PlotSetArray}}
}
\description{
This function takes genomic cooridnates in BED or GFF format,
and extracts the signal from track files (BigWig) and/or
calculates motif density in these regions. Then it computs the statistics
required for average and heatmap plots, and returns the \code{\link{PlotSetArray}}
class.
}
\details{
Function to process genomic signal from tracks and/or motif data,
calculate statistics. This function shoud be used as the entry point to the
seqplots pipeline and fillowed by plotting function(s).
}
\examples{
\dontrun{
ms <- MotifSetup()
ms$addMotif('GAGA')
ms$addMotif('TATA', window=200L, heatmap=TRUE,
            revcomp=TRUE, genome='ce10', name='TATA box')

bed1 <- system.file("extdata", "Transcripts_ce10_chrI_100Kb.bed", package="seqplots")
bed2 <- system.file("extdata", "GSM1208361_chrI_100Kb_PeakCalls.bed", package="seqplots")
getPlotSetArray(ms, c(bed1, bed2), 'ce10')
}
}
\seealso{
Other plotting functions: \code{\link{plotAverage}},
  \code{\link{plotAverage,PlotSetArray-method}},
  \code{\link{plotAverage,PlotSetList-method}},
  \code{\link{plotAverage,PlotSetPair-method}},
  \code{\link{plotAverage,list-method}};
  \code{\link{plotHeatmap}},
  \code{\link{plotHeatmap,PlotSetArray-method}},
  \code{\link{plotHeatmap,PlotSetList-method}},
  \code{\link{plotHeatmap,PlotSetPair-method}},
  \code{\link{plotHeatmap,list-method}}
}

