<<<<<<< 0855384d43d2e83c69bb9ff96f3ed7ead8da615a
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MotifSetup-class.R
=======
% Generated by roxygen2 (4.0.2): do not edit by hand
>>>>>>> Adds rain/ TSCAN/ GOsummaries/ geecc/ seqplots/ systemPipeR/ to the repos.
\docType{class}
\name{MotifSetup-class}
\alias{MotifSetup}
\alias{MotifSetup-class}
\title{MotifSetup Reference Class}
\description{
This class is used to initialize motif plots and mix them with track plots.
}
\details{
<<<<<<< 0855384d43d2e83c69bb9ff96f3ed7ead8da615a
Usage note - the \code{addMotif} and \code{addBigWig} methods can be chained
in following way: \code{MotifSetup()$addMotif("GAGA",
window=200L)$addBigWig("path/to/file.bw")}
=======
Usage note - the \code{addMotif} and \code{addBigWig} methods can be chained in following way:
\code{MotifSetup()$addMotif("GAGA", window=200L)$addBigWig("path/to/file.bw")}
>>>>>>> Adds rain/ TSCAN/ GOsummaries/ geecc/ seqplots/ systemPipeR/ to the repos.
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\section{Methods}{

\describe{
\item{\code{addBigWig(file_path)}}{Adds new BigWig file.}

\item{\code{addMotif(pattern, window = 200L, heatmap = TRUE, revcomp = TRUE,
  genome = NULL, name = pattern)}}{Adds new motif.}

\item{\code{nmotifs()}}{Prints number of motifs in class}
}}
\examples{
#Motifs only
motif1 <- MotifSetup()
motif1$addMotif("GAGA", window=200L)
motif1$addMotif("TATA", window=100L, name="TATA box")

#Motifs and BigWigs
motif2 <- MotifSetup()
motif2$addMotif("GAGA", window=200L)
motif2$addBigWig("path/to/file.bw")

#Chaining commands
<<<<<<< 0855384d43d2e83c69bb9ff96f3ed7ead8da615a
motif3 <- MotifSetup()$addMotif("GAGA")$addBigWig("path/to/file.bw")
=======
motif3 <- MotifSetup()$addMotif("GAGA", window=200L)$addBigWig("path/to/file.bw")
>>>>>>> Adds rain/ TSCAN/ GOsummaries/ geecc/ seqplots/ systemPipeR/ to the repos.
}
\seealso{
Other classes: \code{\link{PlotSetArray}},
  \code{\link{PlotSetArray-class}};
  \code{\link{PlotSetList}},
  \code{\link{PlotSetList-class}};
  \code{\link{PlotSetPair}},
  \code{\link{PlotSetPair-class}}
}

