% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MotifSetup-class}
\alias{MotifSetup}
\alias{MotifSetup-class}
\title{MotifSetup Reference Class}
\usage{
MotifSetup()$addMotif(pattern, window=200L, heatmap=TRUE,
            revcomp=TRUE, genome=NULL, name=pattern)
}
\description{
MotifSetup Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\section{Methods}{

\describe{
\item{\code{addBigWig(file_path)}}{Adds new BigWig file.}

\item{\code{addMotif(pattern, window = 200L, heatmap = TRUE, revcomp = TRUE,
  genome = NULL, name = pattern)}}{Adds new motif.}

\item{\code{nmotifs()}}{Prints number of motifs in class}
}}
\seealso{
Other classes: \code{\link{PlotSetArray}},
  \code{\link{PlotSetArray-class}};
  \code{\link{PlotSetList}},
  \code{\link{PlotSetList-class}};
  \code{\link{PlotSetPair}},
  \code{\link{PlotSetPair-class}}
}

