% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{MotifSetup-class}
\alias{MotifSetup}
\alias{MotifSetup-class}
\title{MotifSetup Reference Class}
\description{
MotifSetup Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\seealso{
Other SeqPlotsClasses: \code{\link{PlotSetArray}},
  \code{\link{PlotSetArray-class}};
  \code{\link{PlotSetList}},
  \code{\link{PlotSetList-class}};
  \code{\link{PlotSetPair}},
  \code{\link{PlotSetPair-class}}
}

