% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifSetup-class.R
\docType{class}
\name{MotifSetup-class}
\alias{MotifSetup}
\alias{MotifSetup-class}
\title{MotifSetup Reference Class}
\description{
This class is used to initialize motif plots and mix them with track plots.
}
\details{
Usage note - the \code{addMotif} and \code{addBigWig} methods can be chained
in following way: \code{MotifSetup()$addMotif("GAGA",
window=200L)$addBigWig("path/to/file.bw")}
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\section{Methods}{

\describe{
\item{\code{addBigWig(file_path)}}{Adds new BigWig file.}

\item{\code{addMotif(pattern, window = 200L, heatmap = TRUE, revcomp = TRUE,
  genome = NULL, name = pattern)}}{Adds new motif.}

\item{\code{nmotifs()}}{Prints number of motifs in class}
}}
\examples{

#Motifs only
motif1 <- MotifSetup()
motif1$addMotif("GAGA", window=200L)
motif1$addMotif("TATA", window=100L, name="TATA box")

#Motifs and BigWigs
motif2 <- MotifSetup()
motif2$addMotif("GAGA", window=200L)
motif2$addBigWig("path/to/file.bw")

#Chaining commands
motif3 <- MotifSetup()$addMotif("GAGA")$addBigWig("path/to/file.bw")
 
}
\seealso{
Other classes: \code{\link{PlotSetArray-class}},
  \code{\link{PlotSetList-class}},
  \code{\link{PlotSetPair-class}}
}

