% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,PlotSetArray,ANY-method}
\alias{plot,PlotSetList,ANY-method}
\alias{plot,PlotSetPair,ANY-method}
\title{Generic plot function for SeqPlots package calsses}
\usage{
plot(x, y, ...)

\S4method{plot}{PlotSetPair,ANY}(x, y, ...)

\S4method{plot}{PlotSetList,ANY}(x, y, ...)

\S4method{plot}{PlotSetArray,ANY}(x, y, ...)
}
\description{
Generic plot function for SeqPlots package calsses
}
\section{Methods (by class)}{
\itemize{
\item \code{x = PlotSetPair,y = ANY}: Method plot for signature 'PlotSetPair'

\item \code{x = PlotSetList,y = ANY}: Method plot for signature 'PlotSetList'

\item \code{x = PlotSetArray,y = ANY}: Method plot for signature 'PlotSetPair'
}}

