% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,PlotSetArray,ANY-method}
\alias{plot,PlotSetList,ANY-method}
\alias{plot,PlotSetPair,ANY-method}
\title{Generic plot function for SeqPlots package calsses}
\usage{
plot(x, y, ...)

\S4method{plot}{PlotSetPair,ANY}(x, what = "a", ...)

\S4method{plot}{PlotSetList,ANY}(x, what = "a", ...)

\S4method{plot}{PlotSetArray,ANY}(x, what = "a", ...)
}
\arguments{
\item{x}{This argument should be one of SeqPlots classes:
\code{\link{PlotSetArray}}, \code{\link{PlotSetList}} or
\code{\link{PlotSetPair}}}

\item{y}{For plotting SeqPlots classes this argument is ignored, used for 
default functionality of \code{\link[graphics]{plot}} function.}

\item{...}{Other parameters controlong the plot, see 
\code{\link{plotAverage}} for avareange plot and \code{\link{plotHeatmap}}
for heatmaps.}

\item{what}{This argument takes a character determining if avareange 
plot (\code{"a"}, default) or heatmap (\code{"h"}) will be plotted.}
}
\value{
Returns \code{NULL} for avareange plot and cluster report \code{data.frame} 
for hetamap - see \code{\link{plotHeatmap}} for details.
}
\description{
Generic plot function for SeqPlots package calsses
}
\section{Methods (by class)}{
\itemize{
\item \code{x = PlotSetPair,y = ANY}: Method plot for signature 'PlotSetPair'

\item \code{x = PlotSetList,y = ANY}: Method plot for signature 'PlotSetList'

\item \code{x = PlotSetArray,y = ANY}: Method plot for signature 'PlotSetPair'
}}
\examples{
#load precalculated PlotSetArrays "plotset1" and "plotset2", usually these 
#objects are the output of getPlotSetArray function
load(system.file("extdata", "precalc_plotset.Rdata", package="seqplots"))

#plot with default values
plot(plotset2) #Average plot
plot(plotset2[1,], what='h') #Heatmap

#setting plot options
plot(plotset2, main='Title', xlab='Relative position [bp]', ylab='Signal', 
     colvec=rainbow(6, 0.7, 0.5), labels=LETTERS, legend_ext=TRUE, 
     legend_ext_pos='topright', legend_pos='topleft', ln.h=9)
     
plot(plotset2[2,], what='h', main="The heatmap", labels=LETTERS,
     ord = c(3,1,2), sortrows = TRUE, clusters = 2, clstmethod = "hclust", 
     cex.main = 20, cex.lab = 12, cex.legend = 12, xlab = "Rel. pos. [bp]", 
     ylab = "Signal", autoscale = FALSE, zmin = 0, zmax = 20, 
     clspace = rev(rainbow(4, 0.7, 0.5)) )

}
\seealso{
\code{\link{getPlotSetArray}}

Other plotting.functions: \code{\link{getPlotSetArray}},
  \code{\link{plotAverage}}, \code{\link{plotHeatmap}}
}

