% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotAverage}
\alias{plotAverage}
\title{external wrapper function, plotting the average plot}
\usage{
plotAverage(plotme, keepratio = FALSE, ord = NULL, labels = NULL, ...)
}
\description{
This function is package internal and should not be executed directly
by users.
}
\keyword{internal}

