% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,PlotSetArray-method}
\alias{plotHeatmap,PlotSetList-method}
\alias{plotHeatmap,PlotSetPair-method}
\alias{plotHeatmap,list-method}
\title{Plot heatmap with optional clustering}
\usage{
plotHeatmap(plotset, main = "", labels = NA, legend = TRUE,
  keepratio = FALSE, ord = 1:length(plotset), scale_signal = "no",
  sortrows = FALSE, clusters = 5L, clstmethod = "kmeans",
  include = rep(TRUE, length(plotset)), ssomt1 = 2L, ssomt2 = 2L,
  title_font_size = 16, ...)

\S4method{plotHeatmap}{list}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, title_font_size = 16, ...)

\S4method{plotHeatmap}{PlotSetPair}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, title_font_size = 16, ...)

\S4method{plotHeatmap}{PlotSetList}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, title_font_size = 16, ...)

\S4method{plotHeatmap}{PlotSetArray}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, title_font_size = 16, ...)
}
\arguments{
\item{tracks}{list or vector of track paths (BigWig) and/or motif
setup structers (as \code{\link{list}})}

\item{features}{vector of feature file paths (BED or GFF)}

\item{bin}{binning window size in base pairs, default 1L}

\item{rm0}{remove 0 from mean/median calculation, default FALSE}

\item{x1}{upsterem distance in base pairs, default 500L}

\item{x2}{downsteram distance in base pairs, default 2000L}

\item{xm}{anchored distance in base pairs, default 1000L}

\item{type}{the type of the calculation, default, 'Point Features'}

\item{add_heatmap}{return the heatmap data,  default FALSE}

\item{cat3}{function, that sends 1st level info to web interface}

\item{cat4}{function, that sends 2nd level info to web interface}

\item{con}{connection to SQlite database storing genome information for files}
}
\value{
nested list: OUT[[FEATURE]][[TRACK/MOTIF]][[VALUE]]
}
\description{
This function plots the heatmap from plot array.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Method for signature plotset='list'

\item \code{PlotSetPair}: Method for signature plotset='PlotSetPair'

\item \code{PlotSetList}: Method for signature plotset='PlotSetList'

\item \code{PlotSetArray}: Method for signature plotset='PlotSetArray'
}}

