% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,PlotSetArray-method}
\alias{plotHeatmap,PlotSetList-method}
\alias{plotHeatmap,PlotSetPair-method}
\alias{plotHeatmap,list-method}
\title{Plot heatmap with optional clustering}
\usage{
plotHeatmap(plotset, ...)

\S4method{plotHeatmap}{list}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  plotScale = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, cex.main = 16, ...)

\S4method{plotHeatmap}{PlotSetPair}(plotset, ...)

\S4method{plotHeatmap}{PlotSetList}(plotset, ...)

\S4method{plotHeatmap}{PlotSetArray}(plotset, ...)
}
\arguments{
\item{plotset}{The dataset to plot - can be \code{\link{PlotSetArray}},
\code{\link{PlotSetList}}, \code{\link{PlotSetPair}} or
properly formated \code{\link[base]{list}}}

\item{main}{The main title of the plot, shown in top-center part of the figure; defaults to NULL (not visible)}

\item{labels}{The character vector giving sub-titles of heatmaps (plotted over the heatmap and below the main title).
The defaults NULL value indicates taht feature/track file names will be used to generate the sub-titles.}

\item{legend}{if TRUE plot the color key}

\item{keepratio}{If TRUE keep 1:1 aspect ratio of the figure; defaults to FALSE}

\item{ord}{The numeric vector determinin the plotting order of experiments.
The heatmap prepesenting feature-track pair with the highest priority will be
plotted leftmost. If NULL (default) the order established in \code{plotset} is used.}

\item{plotScale}{scale the avilable data before ploting, can be "linear" (do not scale, default), "log2" or "zscore"}

\item{sortrows}{If TRUE the rows of heatmap will be sorted by mean value
across all heatmaps, defaults to FALSE}

\item{clusters}{The number of claster for "kmeans" and "hclust", ignored for "ssom", defaults to 5L}

\item{clstmethod}{Determines the heatmap clusering algoritm "kmeans" for k-means (default, see \code{\link[stats]{kmeans}}),
"hclust" (see \code{\link[stats]{hclust}}) for hierarchical clustering, "ssom" for
(super) self organising map (see \code{\link[kohonen]{supersom}}) with trical topology and "none"
of FALSE to turn off the clustering}

\item{include}{The logical vector indicating if given subplot should influence
clustering and sorting, if given element is FALSE the sub-heatmap will be still ploted,
and the order of data rows will be determined vy clusterin/plottin other sub-heatmaps,
defaults to \code{rep(TRUE, length(plotset))}}

\item{ssomt1}{Determins , determins the dimentionality of SOM - number of neurons in 1st dimention,
numbbers of resulting cluters equils ssomt1*ssomt2, defaults to 2L}

\item{ssomt2}{Determins , determins the dimentionality of SOM - number of neurons in 2st dimention,
numbbers of resulting cluters equils ssomt1*ssomt2, defaults to 2L}

\item{cex.main}{Main title font size in points, defaults to 16}

\item{xlab}{label below x-axis}

\item{ylab}{label below y-axis}

\item{cex.axis}{Axis numbers font size in points, defaults to 12}

\item{cex.lab}{Axis labels labes font size in points, Defaults to 12}

\item{cex.legend}{Keys labels font size in points, defaults to 12}

\item{autoscale}{if TRUE the color keys will be autoscaled}

\item{zmin}{global minimum value on color key, ignored if \code{autoscale} is TRUE}

\item{zmax}{global maximum value on color key, ignored if \code{autoscale} is TRUE}

\item{xlim}{the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed and leads to a "reversed axis".
The default value, NULL, indicates that the whole range present in \code{plotset} will be plotted.}

\item{ln.v}{Determins if vertical gide line(s) should be plotted (TRUE) or ommitted (FALSE).
For anchored plots 2 lines indicateing the start and end of anchored distane are plotted.}

\item{s}{The saturation value used to autoscale color key limits, defaults to 0.01}

\item{indi}{If TRUE (defaults) the independent color keys will be plotted below heatmaps,
if FALSE the commmon color key is shown rightmost}

\item{o_min}{vector of length equil to number of sub heatmaps determining minimum
value on color key for each sub plot, if NULL (default) or NA the global settinsg are used,
ignored in \code{indi} is FALSE}

\item{o_max}{vector of length equil to number of sub heatmaps determining maximum
value on color key for each sub plot, if NULL (default) or NA the global settinsg are used,
ignored in \code{indi} is FALSE}

\item{colvec}{The vector of colors used to plot the lines and error estimate fields.
If set value NULL (default) the automatically generated color vaues will be used.
Accpeted values are: vector of any of the three kinds of R color specifications, i.e.,
either a color name (as listed by colors()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb),
or a positive integer i meaning palette()[i]. See \code{\link[grDevices]{col2rgb}}.}

\item{colorspace}{The olorspace of the heatmap, see \code{\link[grDevices]{grDevices}}}

\item{pointsize}{The default font point size to be used for plots. Defaults to 12 (1/72 inch).}
}
\description{
Draw the heatmap plot from \code{\link{PlotSetArray}},
\code{\link{PlotSetList}}, \code{\link{PlotSetPair}} classes or
properly formated \code{\link[base]{list}} (see details) in active graphics window.
Axes and titles, keys and other plot elements are controled by function parameters.
}
\details{
list[[FEATURE]][[TRACK/MOTIF]][[KEY_VALUE]]
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Method for signature \code{\link[base]{list}} with
following format: \code{list[[FEATURE]][[TRACK/MOTIF]][[KEY_VALUE]]}

\item \code{PlotSetPair}: Method for signature \code{\link{PlotSetPair}}

\item \code{PlotSetList}: Method for signature \code{\link{PlotSetList}}

\item \code{PlotSetArray}: Method for signature \code{\link{PlotSetArray}}
}}
\seealso{
Other plotting functions: \code{\link{getPlotSetArray}};
  \code{\link{plotAverage}},
  \code{\link{plotAverage,PlotSetArray-method}},
  \code{\link{plotAverage,PlotSetList-method}},
  \code{\link{plotAverage,PlotSetPair-method}},
  \code{\link{plotAverage,list-method}}
}

