% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,PlotSetArray-method}
\alias{plotHeatmap,PlotSetList-method}
\alias{plotHeatmap,PlotSetPair-method}
\alias{plotHeatmap,list-method}
\title{Plot heatmap with optional clustering}
\usage{
plotHeatmap(plotset, main = "", labels = NA, legend = TRUE,
  keepratio = FALSE, ord = 1:length(plotset), scale_signal = "no",
  sortrows = FALSE, clusters = 5L, clstmethod = "kmeans",
  include = rep(TRUE, length(plotset)), ssomt1 = 2L, ssomt2 = 2L,
  cex.main = 16, ...)

\S4method{plotHeatmap}{list}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, cex.main = 16, ...)

\S4method{plotHeatmap}{PlotSetPair}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, cex.main = 16, ...)

\S4method{plotHeatmap}{PlotSetList}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, cex.main = 16, ...)

\S4method{plotHeatmap}{PlotSetArray}(plotset, main = "", labels = NA,
  legend = TRUE, keepratio = FALSE, ord = 1:length(plotset),
  scale_signal = "no", sortrows = FALSE, clusters = 5L,
  clstmethod = "kmeans", include = rep(TRUE, length(plotset)),
  ssomt1 = 2L, ssomt2 = 2L, cex.main = 16, ...)
}
\arguments{
\item{main}{The main title of the plot, shown in top-center part of the figure; defaults to NULL (not visible)}

\item{labels}{The character vector giving sub-titles of heatmaps (plotted over the heatmap and below the main title).
 The defaults NULL value indicates taht feature/track file names will be used to generate the sub-titles.}

\item{cex.main}{Main title font size in points, defaults to 16}

\item{cex.axis}{Axis numbers font size in points, defaults to 12}

\item{cex.lab}{Axis labels labes font size in points, Defaults to 12}

\item{cex.legend}{Keys labels font size in points, defaults to 12}

\item{xlab}{label below x-axis}

\item{ylab}{label below y-axis}

\item{leg}{if TRUE plot the color key}

\item{autoscale}{if TRUE the color keys will be autoscaled}

\item{zmin}{global minimum value on color key, ignored if \code{autoscale} is TRUE}

\item{zmax}{global maximum value on color key, ignored if \code{autoscale} is TRUE}

\item{xlim}{the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed and leads to a "reversed axis".
The default value, NULL, indicates that the whole range present in \code{plotset} will be plotted.}

\item{ln.v}{Determins if vertical gide line(s) should be plotted (TRUE) or ommitted (FALSE).
For anchored plots 2 lines indicateing the start and end of anchored distane are plotted.}

\item{s}{The saturation value used to autoscale color key limits, defaults to 0.01}

\item{indi}{If TRUE (defaults) the independent color keys will be plotted below heatmaps,
if FALSE the commmon color key is shown rightmost}

\item{o_min}{vector of length equil to number of sub heatmaps determining minimum
value on color key for each sub plot, if NULL (default) or NA the global settinsg are used,
ignored in \code{indi} is FALSE}

\item{o_max}{vector of length equil to number of sub heatmaps determining maximum
value on color key for each sub plot, if NULL (default) or NA the global settinsg are used,
ignored in \code{indi} is FALSE}

\item{colvec}{The vector of colors used to plot the lines and error estimate fields.
If set value NULL (default) the automatically generated color vaues will be used.
Accpeted values are: vector of any of the three kinds of R color specifications, i.e.,
either a color name (as listed by colors()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb),
or a positive integer i meaning palette()[i]. See \code{\link[grDevices]{col2rgb}}.}

\item{colorspace}{The olorspace of the heatmap, see \code{\link[grDevices]{grDevices}}}

\item{pointsize}{The default font point size to be used for plots. Defaults to 12 (1/72 inch).}
}
\description{
This function plots the heatmap from plot array.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Method for signature plotset='list'

\item \code{PlotSetPair}: Method for signature plotset='PlotSetPair'

\item \code{PlotSetList}: Method for signature plotset='PlotSetList'

\item \code{PlotSetArray}: Method for signature plotset='PlotSetArray'
}}

