% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMext}
\alias{plotMext}
\title{Wrapper function, plotting the average plot}
\usage{
plotMext(INPUTS, xlim = NULL, ylim = NULL, main = NULL, xlab = "",
  ylab = "", plotScale = "linear", type = "full",
  error.estimates = TRUE, legend = TRUE, legend_ext = FALSE,
  legend_pos = "topright", legend_ext_pos = "topleft", cex.axis = 14,
  cex.lab = 16, cex.main = 20, cex.legend = 10, ln.v = TRUE,
  ln.h = NULL, colvec = NULL, poinsize = 12, postproc = function() { },
  ...)
}
\arguments{
\item{INPUTS}{list of pre calculated inputs}

\item{main}{an overall title for the plot: see title.}

\item{xlab}{a title for the x axis: see title.}

\item{ylab}{a title for the y axis: see title.}

\item{plotScale}{scale the avilable data before ploting, can be "linear" (do not scale), "log2" or "zscore"}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\description{
This function is package internal and should not be executed directly
by users.
}
\keyword{internal}

