% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{PlotSetArray-class}
\alias{PlotSetArray}
\alias{PlotSetArray-class}
\title{PlotSetArray Reference Class}
\description{
PlotSetArray Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\section{Methods}{

\describe{
\item{\code{nfeatures()}}{Outputs the number of features in the PlotSetArray}

\item{\code{ntracks()}}{Outputs the number of tracks in the PlotSetArray}
}}
\seealso{
Other SeqPlotsClasses: \code{\link{MotifSetup}},
  \code{\link{MotifSetup-class}};
  \code{\link{PlotSetList}},
  \code{\link{PlotSetList-class}};
  \code{\link{PlotSetPair}},
  \code{\link{PlotSetPair-class}}
}

