% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSetArray-class.R
\docType{class}
\name{PlotSetArray-class}
\alias{PlotSetArray}
\alias{PlotSetArray-class}
\title{PlotSetArray Reference Class}
\description{
PlotSetArray Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\section{Methods}{

\describe{
\item{\code{anno(n)}}{Extracts the genomic locations for nth feature as GRanges}

\item{\code{as.array(x, ...)}}{Converts PlotSetArray calss to matrix of PlotSeqPairs}

\item{\code{get(i, j)}}{Subseting method, returns PlotSetArray}

\item{\code{getByID(i)}}{Subseting method, returns PlotSeqList}

\item{\code{getPairs(i)}}{Subseting method, takes pair IDs list, returns PlotSetList}

\item{\code{getRow(i)}}{Subseting method, get row of data as list}

\item{\code{info()}}{Outputs data.frame describing the content of PlotSetList}

\item{\code{nfeatures()}}{Outputs the number of features in the PlotSetArray}

\item{\code{ntracks()}}{Outputs the number of tracks in the PlotSetArray}

\item{\code{pairind()}}{Outputs the list of pair IDs}

\item{\code{plot(...)}}{Plot the PlotSetArray, i.e. all PlotSetPairs within class. 
See \code{\link{plot}} for datails.}

\item{\code{subset(i, j)}}{Subseting method, get PlotSetPair as list}

\item{\code{unlist()}}{Flattens PlotSetArray to PlotSetList}
}}
\section{Subsetting}{
 \code{x} is an object of \code{PlotSetArray} class: 
  \itemize{ \item \code{x[1:2,1:2]} produces \code{\link{PlotSetArray}} with
  2 feature(s) and 2 tracks. \item \code{x[1:2]} produces
  \code{\link{PlotSetList}} with 2 feature/tracks pairs. \item \code{x[[1]]}
  produces single \code{\link{PlotSetPair}}. \item \code{unlist(x)} produces
  \code{\link{PlotSetList}} with all feature/tracks pairs. \item
  \code{x$as.array()} produces the matrix of \code{\link{PlotSetPair}} 
  classes with all feature/tracks pairs. }
}
\seealso{
Other classes: \code{\link{MotifSetup-class}},
  \code{\link{PlotSetList-class}},
  \code{\link{PlotSetPair-class}}
}

