% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSetList-class.R
\docType{class}
\name{PlotSetList-class}
\alias{PlotSetList}
\alias{PlotSetList-class}
\title{PlotSetList Reference Class}
\description{
PlotSetList Reference Class
}
\section{Fields}{

\describe{
\item{\code{data}}{a nested list holding the data}

\item{\code{annotations}}{list of annotations}
}}
\section{Methods}{

\describe{
\item{\code{get(i)}}{Subseting method}

\item{\code{info()}}{Outputs data.frame describing the content of PlotSetList}

\item{\code{npaires()}}{Outputs the number (integer) of PlotSetPairs in the PlotSetList}

\item{\code{plot(what = "a", ...)}}{Plot the PlotSetList, i.e. all PlotSetPairs in the list. 
See \code{\link{plot}} for datails.}
}}
\section{Subsetting}{


\code{x} is an object of \code{PlotSetList} class:
 \itemize{
 \item \code{x[1:2]} produces \code{\link{PlotSetList}} with 2 feature/tracks
 pairs. 
 \item \code{x[[1]]} produces single \code{\link{PlotSetPair}}.
}
}
\seealso{
Other classes: \code{\link{MotifSetup-class}},
  \code{\link{PlotSetArray-class}},
  \code{\link{PlotSetPair-class}}
}

