<<<<<<< 0855384d43d2e83c69bb9ff96f3ed7ead8da615a
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotSetPair-class.R
=======
% Generated by roxygen2 (4.0.2): do not edit by hand
>>>>>>> Adds rain/ TSCAN/ GOsummaries/ geecc/ seqplots/ systemPipeR/ to the repos.
\docType{class}
\name{PlotSetPair-class}
\alias{PlotSetPair}
\alias{PlotSetPair-class}
\title{PlotSetPair Reference Class}
\description{
PlotSetPair Reference Class
}
\section{Fields}{

\describe{
\item{\code{means}}{numeric vector of means}

\item{\code{stderror}}{numeric vector of standard errors}

\item{\code{conint}}{numeric vector of 95\% confidence intervals}

<<<<<<< 0855384d43d2e83c69bb9ff96f3ed7ead8da615a
\item{\code{all_ind}}{numeric vector giving the relative position of the bins in the
genome}

\item{\code{numeric}}{value giving the length of anchored distance, NULL for point
feature plots}
=======
\item{\code{all_ind}}{numeric vector giving the relative position of the bins in the genome}

\item{\code{numeric}}{value giving the length of anchored distance, NULL for point feature plots}
>>>>>>> Adds rain/ TSCAN/ GOsummaries/ geecc/ seqplots/ systemPipeR/ to the repos.

\item{\code{desc}}{character string describing the PlotSetPair}

\item{\code{heatmap}}{numeric matrix used for plotting the heatmap}
}}
\section{Methods}{

\describe{
\item{\code{as.list()}}{Convert to PlotSetPair list.}

\item{\code{plot(what = "a", ...)}}{Plot the PlotSetPair class. See \code{\link{plot}} for datails.}
}}
\seealso{
Other classes: \code{\link{MotifSetup}},
  \code{\link{MotifSetup-class}};
  \code{\link{PlotSetArray}},
  \code{\link{PlotSetArray-class}};
  \code{\link{PlotSetList}},
  \code{\link{PlotSetList-class}}
}

