% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run}
\alias{run}
\title{SeqPlots innitiation}
\usage{
run(root = file.path(path.expand("~"), "SeqPlots_data"), debug = FALSE, ...)
}
\arguments{
\item{root}{the path to data directory, it will be created if not existing}

\item{debug}{run the SeqPlots in debug mode, i.e. with R console embedded in
web interface}

\item{...}{arguments sent to \code{\link[shiny]{runApp}} function}
}
\value{
Normally returns nothing (\code{NULL}), returns an error if one accrued.
Usage messages are shown in R console.
}
\description{
This function initaite SeqPlots and oppens
web browser with graphical user interface (GUI).
}
\details{
The default data directory is "~/SeqPlots_data".
}
\examples{
\dontrun{
run()
}
}
\author{
Przemyslaw Stempor
}
\keyword{seqplots}

