% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\docType{methods}
\name{seqplots-generic}
\alias{[,PlotSetArray,ANY,missing,ANY-method}
\alias{[,PlotSetArray,ANY,vector,ANY-method}
\alias{[,PlotSetList,ANY,ANY,ANY-method}
\alias{[[,PlotSetArray,ANY,ANY-method}
\alias{[[,PlotSetList,ANY,ANY-method}
\alias{seqplots-generic}
\alias{unlist,PlotSetArray-method}
\title{SeqPlots generic methods}
\usage{
\S4method{[}{PlotSetList,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{PlotSetList,ANY,ANY}(x, i, j, ...)

\S4method{[}{PlotSetArray,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{PlotSetArray,ANY,vector,ANY}(x, i, j)

\S4method{[[}{PlotSetArray,ANY,ANY}(x, i, j, ...)

\S4method{unlist}{PlotSetArray}(x)
}
\arguments{
\item{x}{an object of class \code{\link{PlotSetArray}}, 
\code{\link{PlotSetList}} or \code{\link{PlotSetPair}}}

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{see description for \code{i}}

\item{...}{see description for \code{i}}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\value{
If \code{x} is \strong{\code{\link{PlotSetArray}}} class: \itemize{ \item
\code{x[1:2,1:2]} produces \code{\link{PlotSetArray}} with 2 feature(s) and
2 tracks. \item \code{x[1:2]} produces \code{\link{PlotSetList}} with 2
feature/tracks pairs. \item \code{x[[1]]} produces single
\code{\link{PlotSetPair}}. \item \code{unlist(x)} produces
\code{\link{PlotSetList}} with all feature/tracks pairs. \item
\code{x$as.array()} produces the matrix of \code{\link{PlotSetPair}} classes
with all feature/tracks pairs. }

If \code{x} is \strong{\code{\link{PlotSetList}}} class: \itemize{ \item
\code{x[1:2]} produces \code{\link{PlotSetList}} with 2 feature/tracks 
pairs.
\item \code{x[[1]]} produces single \code{\link{PlotSetPair}}. }
}
\description{
Generic operators working with \code{\link{PlotSetArray}}, 
\code{\link{PlotSetList}} and \code{\link{PlotSetPair}} to subset or flatten
the data structure.
}
\seealso{
\code{\link[base]{Extract}}
}

