% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{seqplots}
\alias{seqplots}
\alias{seqplots-package}
\title{SeqPlots - An interactive tool for visualizing track signals and sequence
motif densities along genomic features using average plots and heatmaps.}
\description{
SeqPlots is a web browser tool for plotting average track signals (e.g. read
coverage) and sequence motif densities over user specified genomic features.
The data can be visualized in linear plots with error estimates or as series
of heatmaps that can be sorted and clustered. The software can be run locally
on a desktop or deployed on a server and allows easy data sharing. SeqPlots
pre-calculates and stores binary result matrices, allowing rapid plot
generation. Plots can also be run in batch.
}
\author{
Przemyslaw Stempor
}
\references{
https://github.com/Przemol/seqplots
}
\keyword{CHiP-seq}
\keyword{genomics}
\keyword{plotting}
\keyword{sequencing}

